+++ cmd/tomagnet/main.go
@@ -16,7 +16,7 @@ import (
 )
 
 var (
-	version = "0.2.0"
+	version = "0.3.0"
 	commit  = ""
 	date    = ""
 )